requirements.txt
arabic-reshaper==2.1.3
asgiref==3.3.1
Django==3.1.7
django-cropper-image==1.0.5
easy-thumbnails==2.7.1
future==0.18.2
html5lib==1.1
Pillow==8.1.2
psycopg2==2.8.6
psycopg2-binary==2.8.6
PyPDF2==1.26.0
python-bidi==0.4.2
python-magic-bin==0.4.14
pytz==2021.1
reportlab==3.5.68
six==1.16.0
sqlparse==0.4.1
webencodings==0.5.1
xhtml2pdf==0.2.5
